Optimize deletion process
Problem: If at /tmp/pim-trash is already a file with that name, the actual file won't be moved to that location, but will be removed from self.paths + error message.

Solution: Iterate as long as necessary, until the file in /tmp/pim-trash can be renamed, then move the new/current file to pim-trash.
Pro:
 * No extra module
 * No name collision
 * lesser lines with a for loop
Con:
 * hardcoded to x possible duplicates
 * Dunno how much time it will need, when there is a large number of name duplicates