Move unnecessary stuff out of update_image():
Situation: On every call it was checked if it's an animation, picture or if this stuff was already checked.
Normally that should be checked just once. If move_index() is used. Only time self.pixbufOriginal gets a completly new file.
So the check if the loaded animation is really an animation or just a static image is moved to move_index().
In update_image() just remains the check for a pixbuf or pixbuf animation as they're differenly set.