ToDo: Statusbar
Add a Statusbar to display information.
Per default the statusbar is hidden (due to the quirks) and can be enabled with b or argument -b on startup
The existing keybind for b (back) got renamed to p (previous).

Played a bit with padding, which should minimize the information jumping around in the title and statusbar.

Quirks:
 * My implementation of how the zoom_factor is calculated. If the panel is displayed it still uses the complete window size and not the widget size
 * Information in title bar and statusbar still jumping a bit. And due to the padding the invisible chars make the title bar look a little off