Just jump 1 picture forward, if finding a faulty one
With c5b2f1a I introduced a minimal step of 1, to get actually forward if a defective picture cannot be displayed and the step is = 0.
With some custom keybind and steps of 5, this would also apply for finding a defective picture. But imo there should also just be a step of 1 picture forward.