Fix fullscreen
While removing the class I didn't notice, that the monitor size was elsewhere used (get_zoom_percent()).
Didn't notice this breakage as my wm/compositor decides if something is fullscreen or not.

ToDo:
Take a look at this crazy workaround for solving an issue. Did a first look, but it wasn't promising that this could be removed altogether.